Fix batch protocol alignment with server (Gemini findings)
Fixed 3 critical protocol mismatches identified by Gemini analysis:

1. **Unified BatchResponseDto** - Client now uses single BatchResponseDto
   for all batch endpoints (start/complete/fail/cancel) matching server's
   BatchController.java implementation

2. **Removed fail_batch request body** - Server's failBatch() endpoint
   does not accept request body, removed JSON payload from client

3. **Full DTO structure** - All batch responses now include complete
   BatchResponseDto fields (id, batchId, siteId, status, s3Path,
   uploadedFilesCount, totalSize, hasErrors, startedAt, completedAt)

**Changes:**
- src/batch/dto.rs: Created unified BatchResponseDto matching server
- src/batch/manager.rs: Updated to use BatchResponseDto, removed fail body
- src/service/uploader.rs: Fixed type cast for total_size (i64 → u64)
- tests/common/mod.rs: Updated all mock helpers with full DTO structure
- tests/v2_batch_contract_test.rs: Updated 14 test calls with site_id param
- tests/v2_integration_test.rs: Updated 9 test calls with site_id param

**Test Results:**
✅ All 209 tests passing (156 unit + 40 contract + 13 integration)

+++ src/service/uploader.rs
@@ -269,7 +269,7 @@ impl UploaderService {
             batch_id,
             processed_count: complete_response.uploaded_files_count as usize,
             failed_count,
-            total_size: complete_response.total_size,
+            total_size: complete_response.total_size as u64,
             duration_secs,
         })
     }
